django/requirements.txt
django==3.1.8
django-cleanup==5.1.0
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-image-cropping==1.5.0
django-jazzmin==2.4.4
django-redis==4.12.1
django-simple-history==2.12.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
pillow==8.1.1
pip-chill==1.0.1
pre-commit==2.10.1
python-dotenv==0.15.0
